refactor(cli): address PR review feedback
- Rename plugin folder builtin/sigstore → builtin/oidc (package oidc)
- Restore accidentally removed helm resource repository registration
- Remove MCP knowledge server changes (PR hygiene, dropped via rebase)
- Move OIDC defaulting into GetIDToken (codebase convention)
- Fix goroutine leak: synchronous srv.Shutdown instead of background goroutine
- Remove stale SIGSTORE_ID_TOKEN env var references from help text
- Remove unused blank imports for sigstore v1alpha1
- Improve --verifier-spec flag help: actionable identity constraint guidance

On-behalf-of: Gerald Morrison (SAP) <gerald.morrison@sap.com>
