+++ cli/internal/plugin/builtin/oidc/oidc_plugin.go
@@ -1,4 +1,4 @@
-package sigstore
+package oidc
 
 import (
 	"context"
