+++ cli/internal/plugin/builtin/oidc/oidc_plugin_test.go
@@ -1,4 +1,4 @@
-package sigstore
+package oidc
 
 import (
 	"testing"
